<feature>add/delete Web Services from module palette for SUDSWebServices package</feature> <feature>Package-level context menus in the module palette</feature> <bugfix>added deletion of namespaces in module palette</bugfix>
Package-level context menus can now be created
for the module palette. To use:
First create a function contextMenuName in the package.
It takes the name of the module as parameter and
should return the desired title of the context menu.
Second create a function callContextMenu in the package.
It takes the name of the module and should preform the
desired effect.
See SUDSWebServices package for an example.

Web Services can be removes by right-clicking the
specific Web Service address in the module palette.
Web Services can be added by right-clicking the
"SUDS Web Services" entry in the module palette.

core/modules/package.py:
- Package.has_contextMenuName(self, signature): new - enables custom context menus
- Package.contextMenuName(self, signature): new - enables custom context menus
- Package.has_callContextMenu(self, signature): new - enables custom context menus
- Package.def callContextMenu(self, signature): new - enables custom context menus

gui/module_palette.py:
- QModuleTreeWidget.contextMenuEvent: added call to package context menu
- QNamespaceTreeWidgetItem.get_namespace: check if weak reference has been deleted
- QNamespaceTreeWidgetItem.takeChild: also remove item from index

packages/SUDSWebServices.py:
- contextMenuName: new - returns the correct name for a given signature
- callContextMenu: new - adds/deletes Web Service depending on the signature