<bugfix>Fixed problem where controlflow package would raise duplicate errors</bugfix>
There was an issue where the Fold (base class of Map) module in the
controlflow package would raise ModuleErrors (an object that hosted a
list of ModuleError objects) with two ModuleError objects with the
same module id.  This was problematic because the log controller
expects that a module will not raise more than one error during its
execution.  In addition, we do not need the extra information that the
old exception tried to provide because this is now built-in.

(Since this doesn't change the execution of any workflows, I won't
bump the version number.)

packages/controlflow/fold.py:
  - Fold.compute: don't wrap updateFunctionPort call, let normal error
    handling take care of this