<bugfix>Enable all spreadsheet cells for parameter exploration, even if they are embedded in groups or subworkflows</bugfix> <ticket>#361</ticket>
We build a hierarchy of ids for the virtual cells to deal with
embedded spreadsheet cells.  Note that the numbering of these cells
may not be correct so it may take trial and error to arrange the cells
in the correct fashion.  Note that fixing this would require more
significant updates because a group can contain more than one cell of
the same type and there is no support for showing multiple numbers per
module.  This is filed as a new bug.  Also, note that cells that are
part of a group executed by a control flow will not necessarily have
the desired behavior because there is no way to determine how many
cells will be generated.  However, the straightforward cases should
work, and the logic is recursive so the depth of the cell shouldn't
matter.

Also refactored and simplified the cell layout code to use code from
the controller to ensure that these cells get version numbers and the
validate call succeeds.

core/inspector.py:
  - PipelineInspector.inspect_spreadsheet_cells: make this recursive
    and generate lists of ids instead of single ids to support
    embedded cells
  - PipelineInspector.inspect_ambiguous_modules: label spreadsheet
    cells that are embedded; note that they won't be drawn on the
    pipeline...

gui/paramexplore/pe_tab.py:
  - QParameterExplorationTab.performParameterExploration: pass the
    controller down to the virtual cell positioning code

gui/paramexplore/virtual_cell.py:
  - decodeConfiguration: support nested spreadsheet cells
  - _positionPipelines: support nested spreadsheet cells, simplify the
    code by using the controller's create methods, use a hack to the
    id_scope to get negative ids automatically
  - QVirtualCellWindow.updateVirtualCell: support nested spreadsheet cells
  - QVirtualCellWindow.positionPipeliens: pass controller through