<bugfix>Better error handling for duplicate module identifiers in packages</bugfix>
This fixes a bug reported by Berk Geveci.  Not only do we need to
check that a module isn't identified by two different identifiers but
also that an identifier isn't linked to two different modules.
Created a DuplicateIdentifier exception that is raised when an
identifier is added twice.

core/modules/module_registry.py:
  - DuplicateIdentifier: new exception
  - ModuleRegistry.add_module: check for duplicate identifiers and
    raise an exception if found