<ticket>#406</ticket>
--
When changing version selection in a vistrailView, we enable the "Execute Parameter Exploration" if there's a parameter exploration attached. To keep state consistent (state is loaded when showing the parameter exploration tab) we display the parameter exploration tab just before executing, so the state is loaded.

I think this is a reasonable solution as the user would probably want to see how the parameter exploration was configured when executing it.

To fix the interaction bug, we keep the current cursor mode in the tolbar and upon opening a new vistrail, we reset the cursor of the view to the current state.

gui/vistrail_view.py
  QVistrailView.versionSelected
     Enable the "Execute Parameter Exploration" if there's a parameter exploration attached

 gui/vistrail_toolbar.py
  class QVistrailInteractionToolBar
   __init__()
     new instance variable self.cursorMode
   selectCursorAction()
     setting cursor mode to selection
   cursorToggled()
     updating self.cursorMode according to the current action checked.

gui/builder_window.py
   BuilderWindow
    open_vistrail_without_prompt()
     updating cursor to the current one
    execute_current_exploration()
     switching to parameter exploration view just before executing