<bugfix>Attempt to implement VTK upgrades</bugfix> <ticket>#410</ticket>
I don't know if the code is robust enough to close out the ticket, but
this is an attempt to implement upgrades of ports to the new "_\d+"
style.  Please report any issues you run across while testing this.

Basically, on an initial upgrade request from a VTK module that is
less than the 0.9.3 version, we build a global remap dictionary by
going through all current vtk modules and ports.  If a port now has
the "_\d+" naming, we build a method that when called, attempts to
determine which numbered method to call by checking the signature of
the original port or function.  Note that this cannot be a static
remap (e.g. SetScale -> SetScale_1) because in some cases, SetScale_2
might be correct depending on the signature of the port or function.
If it can't find one, it defaults to the _1 version (maybe it should
raise an error instead?).  This remap dictionary is built once and
stored as a global in the init.py module.

The method is only run when a workflow needs that module and
port/function combination upgraded, but all are defined ahead of time
to take advantage of the remap_module method.  Note that the function
remaps do not generate operations but mirror the original upgrade code
which adds them to the "new module", now passed into the remap method.
The port remaps do, however, generate new operations.

One possible enhancement would be to build the _remap dictionary
incrementally based on the module type.  We only need one remap entry
for each request...

vistrails/core/upgradeworkflow.py:
  - UpgradeWorkflowHandler.replace_generic: pass new_module with
    old_function to the function remap callback

vistrails/packages/vtk/init.py:
  - build_remap: new method
  - handle_module_upgrade_request: new method