<bugfix> Fixed infovis example </bugfix> <bugfix> Fixed NOAA Web services example. NOAA had changed the urls for their services </bugfix>
Removed invalid example offscreen_abstraction.xml

Added running examples in test suite. For now we don't check if the images
are correct but that can be done later. We execute the tagged workflows and
check if there were errors. We are ignoring a few that we don't know that
don't execute such as Isosurface Script in terminator.vt and some intermediate
nodes in other vistrail files.

To run the examples during the test suite just add a -e in the end:
python runtestsuite.py -e

vistrails/tests/runtestsuite.py
  Added optparse support as we are not sending any command line option
to VisTrails anyway.
  So now to run python runtestsuite.py --help to see the supported options.
  Added list of vistrails to be tested with workflow tags to be ignored.