Initial version of a GIS package that uses QGIS.
This is a very simple interface to QGIS that allows one to display
raster and vector layers in the VisTrails spreadsheet.  To load this
package, one needs to install QGIS and its dependencies.  On the Mac,
I had to change some library paths to absolute paths to make sure that
libraries were found.  I also compiled QGIS and PyQwt from source.
Also, note that in the initalize method, there is a hard-coded
reference to the prefix path that will need to be updated for a
different install.

vistrails/packages/qgis/__init__.py:
  - new definition of the QGIS package

vistrails/packages/qgis/init.py:
  - RasterLayer: new module
  - VectorLayer: new module
  - QGISCell: new spreadsheet cell module
  - QGISCellWidget: new spreadsheet cell widget
  - initialize: new method