<feature>New manager for persistent store</feature>
Adds a new manager for the persistent store that allows users to
directly write data to a local directory or delete data from the
persistent store.  This manager can be displayed using the Packages
menu under the Peristence package.  Note that the delete operation
currently only works for **all versions** of a file (ie the top-level
item in the tree).  If you know how to instruct git to forget a
particular version of a file but keep the other versions, I would be
interested in understanding this as the filter-branch magic doesn't
seem to work when constrained to a particular commit.  The manager
shows files and directories in a single view with a type column to
indicate this.

vistrails/packages/persistence/db_utils.py:
  - DatabaseAccess.delete_from_database: new method

vistrails/packages/persistence/init.py:
  - PersistentPath.git_get_dir: run os.makedirs when the output
    directory does not exist
  - PersistentPath.remove_path: new method
  - finalize: delete the configuration widget if it is around
  - menu_items: new method, show a menu item for showing the configuration

vistrails/packages/persistence/widgets.py:
  - PersistentRefModel.__init__: make type the second column, remove
    file_idxs, dir_idxs
  - PersistentRefModel.add_data: no file_idxs, dir_idxs
  - PersistentRefModel.remove_data: new method
  - PersistentRefView.set_visibility: use "type" column directly
    rather than the file_idxs, dir_idxs, this allows id_lists_keys to
    be modified
  - PersistentRefView.get_info: moved to get_info_list
  - PersistentRefView.get_info_list: moved from get_info, allow
    multiple selections
  - PersistentRefViewSearch: new class, factor out view and search box
    as a single widget
  - PersistentPathConfiguration.build_gui: move search/view creation
    to PersistentRefViewSearch
  - PersistentPathConfiguration.search_string: **moved** to
    PersistentRefViewSearch
  - PersistentPathConfiguration.reset_search: **moved** to
    PersistentRefViewSearch
  - PersistentConfiguration: new widget, wraps PersistentRefViewSearch
    with two buttons that allow user to write and delete data from the
    persistent store