<bugfix>Fix issue with caching and persistent files</bugfix>
We define a parameter hasher for the persistent reference that looks
up the version, needed when "most recent" is being used, to determine
if the cached version can be used or not.  In order to do this, we
redefine the "ref" input from a String to a PersistentRef type.  An
upgrade has been added for this translation.  Also fixes a minor issue
where there was an old import left in the code.

vistrails/packages/persistence/__init__.py:
  - version number bumped to 0.2.2

vistrails/packages/persistence/init.py:
  - PersistentRef.get_widget_class: ** removed **
  - PersistentRef.translate_to_python: return None if the eval fails
  - PersistentPath.compute: translate_to_python done automatically now
  - PersistentPath._input_ports: ref : String -> PersistentRef
  - persistent_ref_hasher: new method
  - _modules: add constantSignatureCallable for PersistentRef
  - handle_module_upgrade_request: add upgrade for ref port

vistrails/packages/persistence/widgets.py:
  - PersistentConfiguration.delete: removed old, unused import