<feature>Executable paper editing</feature>
Some modification to add the ability to edit an executable paper
written in LaTeX.  Currently, this is rather preliminary, but we
attempt to parse a LaTeX file, extract the \vistrail commands and push
them into a GUI that resembles the "Embed" GUI.  However, a user can
select different figures, and the idea is that they will be able to
replace figures with new versions.  Then, the modifications can be
written back to the .tex file.

Also fixes a bug with the menubar in Mac OS X.  We define a global
menubar if we're in that OS.

core/publishing/__init__.py:
  - make a valid python module

core/publishing/parse_latex.py:
  - parse_vt_command: new method
  - build_vt_command: new method, inverse of parse
  - parse_latex_file: new method

gui/publishing.py:
  - QLatexFigureItem: new class, subclass QListWidgetItem
  - QLatexAssistant: new class
  - QLatexAssistant.selectSource: callback to select latex source
  - QLatexAssistant.readImage: method to attempt to read pngs into
    interface
  - QLatexAssistant.doCheckLink: method to do checkbox to \vistrail
    links
  - QLatexAssistant.setFigureInfo: method to set interface from latex
  - QLatexAssistant.getFigureInfo: method to set latex from interface
  - QLatexAssistant.figureSelected: callback for list widget selection
    change
  - QLatexAssistant.addFigure: method to add another figure
  - QLatexAssistant.deleteFigure: method to delete a figure
  - QLatexAssistant.saveLatex: callback to write file (hardcoded right now)
  - QLatexAssistant.openVersion: stub
  - QLatexAssistant.useCurrent: stub
  - QVersionEmbed: ** moved ** from version_prop.py, some layout changes

gui/version_prop.py:
  - QVersionProp: remove references to QVersionEmbed

gui/vistrails_window.py:
  - QVistrailsWindow.init_palettes: add QVersionEmbed/QLatexAssistant
  - QVistrailsWindow._global_menubar: ** added **
  - QVistrailsWindow.createActions: use _global_menubar if we're in
    Mac OS X