Reverted palette's title bar to default QDockWidget.titleBarWidget Added toolbar to palette with a radio button to work as a pin button.
--
I was not completely satisfied with the behavior of the custom title bar for
the palettes. Tommy also noticed that they looked werid in Linux. So I
reverted to the default title bar and added a toolbar to the palette
in QToolWindow (the radio button appears right below the close button).
It doesn't look as nice as a button on the title bar, but I think in this
case the dragging/switching floating status is more important than the
appearance.

--
gui/common_widgets.py
  class QToolWindow:
    __init__()
      added a main window as the widget of QDockWidget so a tollbar can be
      used. The original widget is now set as the central widget of the window.
      removed title_bar_widget member

    new methods:
      createToolBar()
      setDefaultPinStatus()

    removed methods:
      toggleTopLevel()
      changeTitleBarToDefault()
      changeTitleBar()

   class QToolWindowInterface:
     toolWindow()
       now checks if self is the central widget of the window in dockwidget

   Removed class QToolWindowTitleBar and QToolWindowTitleButton