Removed detach button from tabs and using mouse double-click event to detach tabs
--
For some unknown reason, the detach button was working only on Mac. To avoid
the problems that were happening in the other systems, we removed the detach
icon and the user can double-click on a tab to detach it.
There will be a toolTip on the detachable tabs.

gui/common_widgets.py
  added new class QMouseTabBar that emits a signal on mouse double click events

  gui/vistrail_view.py
    class QVistrailView
        replaced regular QTabBar by QMouseTabBar
	removed button_to_tab_idx map
	removed setTabDetachable method
	removed setAllTabsDetachable method
        added new method isTabDetachable
	added new method updateTabsToolTip
	added new method tabDoubleClicked
       in remove_view_by_index()
          updates tabs tooltips after removing the tab
       in view_changed()
          updates current tab tooltip