Fixed button sizes on Palettes. Also adjusted the margins and spaces in the layouts of some palettes so they look nicer.
--
The problem is that setting a minimum height 30 for QPushButton was not enough
for making it look nice on Mac. I checked on Designer and I couldn't reproduce
this problem. The butons always looked good when docking/undocking the palette.
It seems it sets the minimum height behind the scenes somehow. The value for
minimum height there is 32 but that is too much for Windows, even 30 looks
already bad on Windows. So I created a QDockPushButton that sets the minimum
height only on Mac. This should solve the problem for Macs and Windows. Need
to check how it will look on Linux.

--
gui/common_widgets.py

 Added new class QDockPushButton

gui/version_prop.py
 QVersionProp.__init__()
   adjusted margins and spaces in the layouts

 QVersionThumbs.__init__()
   adjusted margins and spaces in the layouts

gui/publishing.py

 QLatexAssistant.__init__()
   replaced QtGui.QPushButton by QDockPushButton
   removed calls to setMinimumSize in pushbuttons

gui/module_info.py
 QModuleInfo.build_widget()
   setting minimum height of lineedit to 22
   replaced QtGui.QPushButton by QDockPushButton
   removed calls to setMinimumSize in pushbuttons
   adjusted margins and spaces in the layouts

gui/debug.py
  DebugView.__init__()
    replaced QtGui.QPushButton by QDockPushButton
    setting fixed width to 125 because 120 was not wide enough on Mac.