<feature>Simplify visual diff and improve parameter changes display</feature> <ticket>#452</ticket>
Update the diff logic to check when parameter settings are equal so
they are not classified as "changes".  Note that this can be
problematic for parameters that are set multiple times, but in general
should be better.  Also remove the added "Matched" label; such modules
are now classified as "Shared".  The differentiation only helped those
who wished to better understand the algorithm; users were likely
confused.

db/services/vistrail.py:
  - getParameterChanges: if using heuristic matching, don't add a
    change if the match is good
  - getWorkflowDiff: check heuristic pairs for parameter changes, and
    if there are none, classified them as matched

gui/vis_diff.py:
  - QLegendWindow.__init__: remove Matched classification
  - QDiffView.set_diff: merge Matched classification into Shared