<ticket>#456</ticket>
This fixes the problem of not being able to copy/paste on a detached view.

As the menu actions' state is global, when switching the focus to a detached view,
the parent QVistrailWindow will have the actions for that view activated, even though the
visible tab is not of the same type. This means that you might see the execute button
activated because you have the focus on a detached pipeline view even though the main
window displays a history view. When trying to press the execute button, the button will be
disabled immediately because that window got the focus. So to execute a detached pipeline, you
have to use the menu Workflow > Execute.

We will probably want to add a toolbar for the detached views with the most common actions.

--

gui/vistrails_window.py
  class QVistrailsWindow
    set_action_links()
    set_action_defaults()
      if the object is a detached window, we will consider the window as
      the window of the object's vistrail_view

   applicationFocusChanged()
      making sure the vistrail_view updates the current tab when the focus
      changes from/to a detached view

gui/vistrail_view.py
   class QVistrailView
      get_gurrent_tab()
         reorderd test logic so if the current window is not a detached tab,
          it will return the current tab of this window.

      view_changed()
          making sure to also consider the current_tab as a detached tab.

      showCurrentViewPalettes()
          making sure to also consider the current_tab as a detached tab.