Bug fixes  1) Fixed problem in which VisTrails would try to create a vistrail     when quitting  2) Fixed problem in which VisTrails would crash when closing a vistrail     soon after changing a parameter on a Mac
--

In 1) using QApplication.closingDown() was not working because
 closing_all_vistrails() is done before the application gets to the
 closingDown stage. So we have to set a variable and keep track of
 it ourselves.

In 2) it seems a Qt bug. I left a comment in the code explaining the problem,
but it happens on a Mac if you set a parameter and soon after close the
vistrail (with cmd+W). Clearing the focus of the widget fixes the problem.

gui/vistrails_window.py
  class QVistrailsWindow:
   __init__()
     added _is_quitting member to keep track when the application was requested
     to close

   close_vistrail()
     replaced QtCore.QCoreApplication.closingDown() with self._is_quitting

   quit()
     set self._is_quitting = True
     if the user cancels the process, set it to False

gui/ports_pane.py
   PortsList.update_module()
      clearing the focus if a child widget has the focus before clearing