<bugfix>Changes to subworkflow identification</bugfix> <ticket>#446</ticket>
In order to address serious issues with subworkflow editing, the
subworkflow identification scheme has been modified to encode more
information.  The previous code required that we modify past
subworkflow information in vistrails that use the subworkflows which
is a bad idea for provenance information.  Instead of renaming
subworkflows when they are saved to vistrail files, we now encode
namespaces when a subworkflow is modified.  At the same time, we
maintain all previous namespaces for a subworkflow.  Thus, each
subworkflow can have a collection of annotations that adhere to the
regex __abstraction_uuid(_%d)?__.  For example, a subworkflow might
have the annotations:

<annotation id="1" key="__abstraction_uuid__" value="bcb5c58c-b959-11e0-a88d-0023dfde3d57"/>
<annotation id="2" key="__abstraction_uuid_2__" value="aefcfc26-ba08-11e0-a2af-0023dfde3d57"/>
...

The current namespace of a subworkflow is the last uuid (the highest
numerical value).  However, each namespace remains valid so an older
subworkflow reference can be loaded.  Thus, if a vistrail contains a
subworkflow reference:

 <abstraction cache="1" id="4" internalVersion="1" name="str_test15" namespace="bcb5c58c-b959-11e0-a88d-0023dfde3d57" package="local.abstractions" version="1.6" />

it can be loaded from the registry using the old namespace.  Note that
this does require sidestepping the DuplicateModule error so the
_module_key_map in the registry is no longer valid for subworkflows.
In other words, we can have multiple descriptors pointing to the same
module.  However, most code should not be using get_descriptor on
subworkflows anyway.  Because we have a unique annotation for each
saved version, we can now also determine which subworkflows contain
the other to reduce duplication.

There are likely some items that will need to be tweaked, but this
commit should resolve some of the major errors.

core/modules/module_registry.py:
  - ModuleRegistry.add_module: if we're adding a subworkflow, don't
    raise the DuplicateModule exception

core/modules/sub_module.py:
  - get_abs_namespace_info: new method, get namespace info for subworkflows
  - get_all_abs_namespaces: new method, convenience
  - get_cur_abs_namesapce: new method, convenience
  - get_next_abs_annotation_key: new method
  - get_cur_abs_annotation_key: new method
  - new_abstraction: use get_cur_abs_namespace to obtain uuid

core/vistrail/controller.py:
  - VistrailController.create_module_from_descriptor_static: make sure
    to use the latest namespace for a subworkflow
  - VistrailController.create_abstraction: new path to obtain
    namespace
  - VistrailController.create_abstraction_from_group: new path to
    obtain namespace
  - VistrailController.add_abstraction_to_registry: load all
    namespaces for an abstraction so old uses will work
  - VistrailController.load_abstraction: new path to obtain namespace
  - VistrailController.unload_abstractions: simplify code, unload all
    namespaces
  - VistrailController.save_abstraction: use new path to set uuid for
    abstraction
  - VistrailController.upgrade_abstraction_module: re-read the
    vistrail and force use of current namespace
  - VistrailController.import_abstraction: new path to set namespace
  - VistrailController.write_abstractions: new method, new logic to
    determine how to write abstractions out
  - VistrailController.write_vistrail: factored out abstraction
    writes, do not do an unload of abstractions here