<bugfix>Fix issue with package subworkflows and	namespaces</bugfix>
Matthias found an issue with subworkflows where the namespace was
assigned incorrectly.  This was due to a change to local subworkflows
where the namespace needed to be checked on the addition of a
subworkflow to a workflow.  We now only assign the namespace for local
subworkflows, not for those in packages.

core/vistrail/controller.py:
  - VistrailController.create_module_from_descriptor_static: namespace
    is based on descriptor when subworkflow is not from
    local.abstractions