<bugfix>Fixed an issue when loading a vistrail that references different versions of a subworkflow that share common ancestors</bugfix>
Now that we load all versions of a subworkflow so that old versions
can reference the old namespaces, if we have a "branch" that created
two subworkflow files, we should not load their common ancestors more
than once.  This can be accomplished by checking the registry before
adding.  Because we are using uuids, if we see the same namespace, we
have the same underlying workflow at that point.

core/vistrail/controller.py:
  - VistrailController.add_abstraction_to_registry: check if the
    subworkflow already exists in the registry