<bugfix>Fixed issues with multiple versions of subworkflows appearing in the module palette</bugfix>
Because we are now loading multiple versions of a subworkflow when
reading it, we need to be careful about which version appears in the
module palette.  We now set hide_descriptor to False only when the
version being loaded is the current one.  In addition, the
update_module call uses this current version.

This also fixes an issue where upon saving a subworkflow after
editing, we forced it to appear in the module palette.  We should only
do an update; if a subworkflow already appeared in the palette, we
will swap in the new version, but otherwise, we will hide the
descriptor.

core/vistrail/controller.py:
  - VistrailController.add_abstraction_to_registry: track current
    namespace and only display one version in the palette, also keep
    track of which version was requested and return it
  - VistrailController.write_vistrail: call load_abstraction with
    is_global set to False so we only replace a subworkflow that is
    already in the module palette