<bugfix>Fixed issue where subworkflow edits could result in one-to-many port mappings.</bugfix> <ticket>#460</ticket>
Add a check in write_vistrail to check port names for subworkflows.
The problem usually shows up when a user adds a new port and connects
it to a port with a name that is already connected to another port.
Because VisTrails uses the connected port names, both ports would have
the same name and appear as a single port in the updated subworkflow
module.  The new method registers port names that have already been
set and assigns names to new ports to avoid the overlap.  Note that
user-defined overlaps are not corrected; if a user manually assigns
the same name to two different ports, we raise an exception upon the
user trying to save.  Otherwise, we would be guessing at which should
keep the assigned name, and this would cause problems for upgrades.
Also changed the error for save_vistrail to be more explanatory.

core/vistrail/controller.py:
  - VistrailController.check_subpipeline_port_names: new method
  - VistrailController.write_vistrail: do check of port names before
    saving

gui/vistrail_view.py:
  - VistrailView.save_vistrail: display a more helpful error message