<feature>Changed "Keep" button to "Tag" in the Mashups View</feature>
--

As discussed in the VisTrails Dev meeting, we thought better to
change the name of the "Keep" button because it didn't actually
keep it if you didn't save the vistrail. The now "Tag" button will
just give a name to the mashup.

gui/mashups/mashup_view.py

In QMashupView.createActions()
    Renamed the Keep button to Tag and changed its toolTip accordingly.