<ticket>#491</ticket> <bugfix>Opening vtl files: igonring showSpreadsheetOnly if execute is false</bugfix> <bugfix>Starting VisTrails by double-clicking a vistrail file was not zooming the current pipeline view</bugfix>
- Displaying [invalid view] warning only for unknown window types.

When extracting the workflow/vistrail from a .vtl file and a file with the same name exists and it is different from the workflow/vistrail we are extracting, we will ask the user if we can overwrite it. If not, the user will be asked to give another name.

--

gui/vistrails_window.py
  In QVistrailsWindow.get_current_view()
      In case the window is None or it is a dialog box, we omit the warning

gui/extras/core/db/locator.py
  Added new function ask_to_overwrite_file()

gui/application.py
  class VistrailsApplicationSingleton:
      Refactored code to display builder window into a method showBuilderWindow
      interactiveMode()
          using showBuilderWindow

  class VisTrailsApplicationInterface:
     process_interactive_input()
        making sure the builder window is shown before opening the vistrail.

core/db/locator.py
  In FileLocator.from_link_file()
      ignoring showSpreadsheetOnly when execute is False
      asking for user's input when a file already exists.