<bugfix>Add upgrade for PythonSource to fix issue with 'self' port</bugfix> <ticket>#492</ticket>
A fix for the issue reported where the "self" port of PythonSource is
expected to be of type "PythonSource" instead of "Module".  I added an
upgrade that just changes it to the most recent port type, but I do
wonder exactly why the change was made by Fernando (in 2009).

Also found one more place where an upgrade message should be at the
log level.

core/modules/basic_modules.py:
  - handle_module_upgrade_request: add remap for PythonSource

core/upgradeworkflow.py
  - UpgradeWorkflowHandler.dispatch_request: change level of info
    about package not handling upgrade request