<bugfix>Fixed issue with ports that may have multiple functions associated with them (e.g. in the vtk package)</bugfix> <ticket>#495</ticket>
This makes sure that functions are not replaced when there have been
multiple functions added.  Note that we are still missing the ability
for package developers to specify the cardinality of a port.  This
would allow us to enable/disable the + button for ports based on their
cardinality.

gui/ports_pane.py:
  - PortsList.update_method: pass should_replace on to controller
    based on whether there is an associated function id or not

gui/vistrail_controller.py:
  - VistrailController.update_function: add new should_replace
    parameter which is passed through to update_function_ops