<ticket>#494</ticket>
--

The Module Configuration window sometimes can have its title
changed to include a '*' indicating that there are changes that
are not persisted. When the window was closed after a change and
before saving/resetting, the title was never reset and it was
inconsistent with the title that QVistrailsWindow expected (without the *).
The title is reset everytime the module changes.

gui/module_configuration.py
In QModuleConfiguration.updateModule()
   also reset title to 'Module Configuration'