<feature>Expand search results to include all open vistrails</feature>
A bunch of changes to allow searches to include all open vistrails
(included unsaved vistrails).  Also the navigation of search results
is now self-contained so matching versions and modules are shown in
the query view; there is an "Edit" button that opens the corresponding
entity in "normal" editing mode.  Note that the global results are
currently shown in the workspace window.  The unsaved results are tied
to temprorary entities in the core.collection.Collection instace, and
links to their windows are maintained by the workspace palette.

The inclusion of global results leads to some odd paths through views.
Because a query view exists for each vistrail, a search in one
vistrail that returns global results can lead to a user viewing the
result from another vistrail (digging down to a specific workflow, for
example).  At the time of selection, we switch the view to the
selected vistrail, and we also pass the contents of the search over to
that vistrail's query view.  However, the query pipeline and/or search
strings are not passed through, so clearing the search in the new
query view leads to an empty (or another) query pipeline or string.  I
think there probably needs to be single global query view, but pulling
that out at this point would take some time.

Also, I fixed an annoying bug with ensureVistrail which didn't reset
the tab states.

core/collection/__init__.py:
  - Collection.__init__: add temp_entities dict
  - Collection.get_current_entities: new method
  - Collection.add_temp_entity: new method
  - Collection.is_temp_entity: new method
  - Collection.delete_entity: make temp_entity aware
  - Collection.fromUrl: return None if the url cannot be found

core/collection/entity.py:
  - Entity.__init__: add is_open field

core/query/multiple.py:
  - MultipleSearch: new search class, takes a dictionary {entity:
    versions_to_check} so it can search multiple vistrails

gui/collection/workspace.py:
  - QWorkspaceWindow.gotoSearch: set searchMode on list
  - QWorkspaceWindow.updateSearchResults: pass search object through
  - QVistrailList.__init__: init searchMode and search obj
  - QVistrailList.update_search_results: set search object so it can
    be used to open search results
  - QVistrailList.search_result_selected: new method
  - QVistrailList.item_selected: default entity to None, use _window
    attribute to locate views for unsaved entities, switch execution
    based on searchMode
  - QVistrailList.add_vt_window: create entity if needed and set
    is_open and _window
  - QVistrailList.remove_vt_window: reset entity attributes, don't add
    to closed list if it was temporary
  - QVistrailList.setSelected: refactor selection code so it works for
    searches and open items

gui/query_view.py:
  - QueryController.search_level: default search level
  - QueryController.set_search: old method renamed to run_search, new
    method just sets search objects
  - QueryController.run_search: moved from set_search: check search
    level to determine if recompute is necessary, use MultipleSearch
    for all types of search now, pass search as part of search_changed
    notification
  - QueryController.goto_edit: new method
  - QueryController.update_version_tree: let set_search set
    controllers
  - QueryController.show_vistrail_matches: set level and remove
    controller sets
  - QueryController.show_workflow_matches: set level and remove
    controller sets
  - QueryController.show_global_matches: new method
  - QQueryPipelineView.execute: call run_search instead of set_search
  - QQueryResultGlobalView.__init__: new message and align correctly
  - QQueryBox.build_widget: change text to "Open Vistrails", add edit button
  - QQueryBox.doEdit: new method
  - QQueryBox.executesearch: call run_search instead of set_search
  - QQueryView.update_controller: new method
  - QQueryView.set_to_search_mode: set edit button to disabled
  - QQueryView.set_to_result_mode: set edit button to enabled in some
    cases
  - QQueryView.result_version_selected: set current vistrail on the
    search (needed for match/matchModule)

gui/vistrail_view.py:
  - QVistrailView.query_version_selected: new method

gui/vistrails_window.py:
  - QVistrailsWindow.ensureVistrail: call change_view instead of
    setCurrentWidget to ensure tab states are set correctly