<feature>CLTools - package for wrapping command line tools</feature>
USAGE:

Create a file {modulename}.conf in the directory .vistrails/CLTools/
Using the following JSON syntax:

ROOT is a dict with the following possible keys:
* command (required) - value is the command to execute like "cat" or "/home/tommy/cat"
* stdin - handle stdin - value is a 2-list ["port name", CLASS]
* stdout - handle stdout - value is a 2-list ["port name", CLASS]
* stderr - handle stdout - value is a 2-list ["port name", CLASS]
* args - list of ordered arguments that can either be constants, inputs, or outputs. See ARG.
ARG is a 4-list containing [TYPE, "port name", KLASS, OPTIONDICT]
CLASS can either be "File" or "String". Unknown types are handled as "String".
TYPE is one of:
* input - create input port for this arg
* output - create output port for this arg
* flag - use "port name" directly as a constant string
OPTIONDICT is a dict containing argument options. recognized options are:
"list": "" - treat input as a list of objects of the specified CLASS. This enables variable number of inputs.
"flag": "name" - Append name as a constant before the specified argument. If it is also a list it is appended before each item
"prefix": "name" - Append name as a prefix to the final argument. If it is also a list it is appended to each item.

EXAMPLE:

wrap the command "cat" that takes 2 files as input named "first" and "second". Also take a list of files as input named "rest".
Catch stdout as file, name it "combined".
Catch stderr as string, name it "stderr".

{   "command": "cat",
    "args": [["input", "first", "File", {}],
             ["input", "second", "File", {}],
             ["input", "rest", "File", {"list":""}]],
    "stdout": ["combined", "File"],
    "stderr": ["stderr", "String"]
}

Save as /home/tommy/.vistrails/CLTools/cat.conf
Reload CLTools package. Test the new module.

CLTools/__init__.py: *new - declares the CLTools package

CLTools/init.py: *new
- CLTools: *new - declares the abstract CLTools module that is baseclass to all wrapped tools
- add_tool: reads a tool description from the supplied path and creates a runnable vistrails module
- initialize: load all tool descriptions in the .vistrails/CLTools folder.
- finalize: