Fixed problem in nightly build script
--
 The problem in the nightly build script was that the tarball
was not being uploaded to sourceforge because no revisions were
being detected in that day. The script used to be run before
midnight and now runs at 3AM, so now we need to check for
changes made the day before.

Also updated released revision in the master branch so the
list of changes in the nightly built is correctly computed.

vistrails/core/system/__init__.py
  release is now 1b88c3949efd

dist/source/make-vistrails-src/nightly.py
  DATETIME_START is now being computed as the day before the script is run.